[server] Move websocket-connection-manager.ts
